php-to-yaml/pyproject.toml
[tool.poetry]
name = "php-to-yaml"
version = "0.1.0"
description = "Converts the rules contained within the PHP files into YAML"
authors = ["Your Name <you@example.com>"]
license = "GNU"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.13.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
